geändert vom Long an Int bei id

+++ src/main/java/com/example/demo/model/User.java
@@ -1,15 +1,11 @@
 package com.example.demo.model;
 
 
-import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 import java.util.Date;
+
 
 @Entity
 @Getter
@@ -18,7 +14,7 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     private Date birthdate;
